--- backend/src/main/java/com/shinhan/shbhack/ijoa/api/controller/member/AuthController.java
@@ -1,17 +1,0 @@
-package com.shinhan.shbhack.ijoa.api.controller.member;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@Slf4j
-@RestController
-@RequestMapping("/api1/auth")
-@RequiredArgsConstructor
-public class AuthController {
-
-
-
-
-}
